[refactor] acq stream: split SPARC test cases
One file per type of SPARC. This is mostly to help with the default 30
min timeout during CI.
Running the test for a single type of SPARC can easily take 20 min.
Because of this, the test cases where always stopped before the entire
execution was finished, possibly shadowing some failures.
